[WIP] improve the calculation of processed rows
The new calculation takes the value of Actual Loops into consideration.
However, since this seemingly is not always necessary, this
overestimates the number of actually processed rows from time to time. A
safer alternative is to calculate the number of processed rows directly
based on the live database (if that is possible).