[WIP] UES preparation: attribute parsing for predicates
This enables MospPredicates to provide full AttributeRefs for the
attributes that are referenced in their predicate. In order to do so, an
alias map is introduced, which provides for each alias the full TableRef
instance (e.g. n would map to TableRef("name", "n"), kt to
TableRef("kind_type", "kt") and so on).