[WIP] UES preparation: Add support for compound filter predicates
This introduces a new kind of MospPredicate, namely the
CompoundMospFilterPredicate. In contrast to its counterpart, instances
of this class can represent logical combinations of filter predicates.
E.g. the WHERE clause `WHERE r.a = s.b AND (r.c = 42 OR r.c = 24)` can
now be parsed completely and correctly (hopefully..). Note that a
compound filter predicate really is just that. Join predicates are not
supported (as in `WHERE (r.a = s.b) OR (r.a = s.c)`).