Support non-SPJ queries in UES optimization
If queries contain GROUP BY, ORDER BY or HAVING clauses, as well as
aggregations as part of their projection, optimization still works.

This is achieved by transforming the query into a SPJ "equivalent" which
is suitable for optimization, and reconstructing the original query
afterwards. However, the query under optimization is not equivalent from
a correctness point of view. Instead, it will simply have all additional
clauses and aggregations removed. During reconstruction, these will then
be added back in. This approach also allows queries that have no table
aliases to be optimized (following the same strategy as before, i.e.
adding temporary aliases). At some point in the (near) future, this may
also allow queries with explicit joins to be supported.

As part of these changes, function calls that accessed the db schema but
did not receive a correct db schema instance from the callee have also
been fixed. This is necessary if the db schema is initialized from data
other than the current .psycopg_connection file contents (such as when
testing JOB workload and SSB workload at the same time..).