Change UES query preparation behavior regarding aliases
Previously, all generated aliases would be dropped from the final query.
This seemed reasonable since the query should not be modified
unnecessarily. However, most of the logic in mosp.MospQuery also relies
on the existence of table aliases and fully-qualified attributes (which
is a design issue on its own, originating from the nice and
fully-qualified JOB world). Therefore, dropping all attributes would
make it impossible to analyze the optimized query with the current
MospQuery implementation. Therefore, the default behavior is now to
keep the generated attributes/table renamings. If this is not desired,
passing the corresponding function argument will switch to the previous
behavior.