Switch PG params depending on PG version
This initial version implements functionality to
a) extract the Postgres version from the database via the DBSchema
interface
b) compare versions conveniently, e.g. as `Version("14.2") > 12`

Based on this functionality, the ues-generator.py script can now
dynamically use the `SET enable_memoize = 'off';` setting if Postgres 13
or later is used (which is when the parameter was introduced). However,
as of now this functionality is limited to live execution of single
optimized queries and most importantly it is left unused during workload
execution, e.g. in experiment-runner.py. This is due to these scripts
using a large number of legacy features (e.g. using psycopg connections
directly instead of through DBSchema). A future refactoring of workload
execution will enable these scripts to utilize version-dependent
parameters as well.