Change some DBSchema initialization patterns
Instead of using the default DBSchema instance as a default parameter in
function definitions, the DBSchema will now be initialized first thing
in the function body. This has two nice side effects:
1) if DB functionality is not necessary, no DBSchema conncetion will be
created and hence no running database is required
2) if the default DBSchema changes within the application, it will be
automatically updated in the function bodies (as long as no older
DBSchema is supplied as a function argument)

Eventually, all usage of DBSchema should be refactored to this pattern.