Change query generation for postgres to use Leading instead of JOIN
This modifies the generation of optimized queries to no longer transform
the query using the explicit JOIN ON syntax in combination with
join_collapse_limit = 1. Instead, the join order is enfored via the
Leading() hint of pg_hint_plan. This has two practical implications:
first up, the join tree is also used to infer join directions. Each join
node is interpreted as follows: the child with smaller bound becomes the
outer relation and the larger bound child becomes the inner relation. If
no bounds are set, the left child becomes the outer relation by default.
As a second consequence, this means that operator assignments have to
considered while generating the Leading() hint. This is because Postgres
is inconsistent (or rather sometimes unintuitive) with the assignment of inner/outer relations. More specifically, for a hash join the inner relation is the one being hashed and the outer relation is probed. PostBOUND thinks of hash joins the other way around: the inner relation is probed whereas the outer relation is used to construct the hash table.