Add new optimization tests based on EXPLAIN
These tests only ensure that the generated optimized query still
constitutes a valid SQL query by passing it as an EXPLAIN query to the
database system. The tests do not ensure that the optimized query still
produces the same result set as the original query. This is still left
to the former test methods (now called test_result_set_equivalence).