Large refactoring: introduce physical query plan and logical join tree
This also changes the behaviour of the statistics container to only use
explicit updates. This makes the overall code design (especially the
implementation of general join trees) much simpler while making the
optimizer code only slightly more complex (by adding the explicit
call to the statistics update). In the end, the statistics container
will probably become part of the "private" UES implementation anyway
instead of being a public utility for anyone to use. Its contribution is
simply too small and too much tailored to UES for that.

Missing bits of documentation will be added later on.