More fine-grained error handling for Stack benchmark test errors
This is necessary because UES apparently chooses pretty bad plans. These
plans lead to extensive spilling to disk, which in turn results in the
query failing with a disk full error. This is not really an error from
the test's POV and we therefore skip such queries.

A likely cause for this situation could be the usage of hash joins and
hash tables that become larger than memory. The combination of forced
hash joins with bad hash join assignments (inner vs outer relation,
which is controlled by the native PG optimizer) can be pretty bad for
performance (along with a bad join ordering).