Make Join trees no longer iterable.
The combination of `Container[TableRef]` and `Iterable[Node]` violated an
invariant specified in PEP 3119: Sized and Iterable objects have to
fulfill `sum(1 for x in c) == len(c)`.
However, since the size was based on the number of base tables in the
join tree and the iteration happend in terms of intermediate join nodes,
those two numbers did not add up.
This lead to problems in third-party code that assumed the invariant (as
it reasonable should) and tried to iterate like so `(next(c) for __ in
range(len(c))`.
Since the iteration is directly based on the `join_sequence()` method,
and this method still exists, the implementation of the Iterable
interface was just a convenience measure anyway and dropping it does not
hurt anybody.