Make sure to return a copy of the join graph in QueryPredicates
This prevents modifications of the join graph in client code leaking to
other usages of the join graph due to the usage of the functools cache.

At the same time, this commit expands the information contained in the
join graph to also include the join predicates as annotations on the
respective edges.