Improve __str__ method for join trees
This makes the internal structure of the join tree more transparent and
no longer prints join trees in a depth-first manner. This was a pretty
bad idead because different join trees (e.g. left-deep and right-deep)
looked similar in the output despite their structure being very
different.

This also makes the representation of base tables shorter if an alias is
available.