Improve performance of QueryPredicates.joins_between method
This changes the default implementation of said method to operate on
pre-computed data nearly all the way through. At the same time, this
means that when constructing a QueryPredicates object, slightly more
work has to be done. I argue that this is OK, since QueryPredicate
objects are only constructed on-demand and will cached for each query.
Therefore, they rarely need to be built.

Just to be save, this commit includes the former (now legacy) behavior
for computing the joins_between method as well as another graph-based
implementation in case we need to switch this around later.