Fix implementation of PhysicalQueryPlan.plan_hash
The previous implementation used functools.cache to store the hash
value once it had been calculated. However, this decorator is based
on the actual hash value as calculated by __hash__. This hash value
is only calcualted based on the join order. As a consequence,
physical query plans with different operators but the same join
order still returned the same hash.