Fix Postgres GeQO management and restore behaviour
The old implementation only obtained the current GeQO state once during
creation of the Postgres interface and never updated this state. If some
external component modified the state in the meantime, these changes
would not be reflected in the current Postgres interface. Therefore,
after executing a query, the old GeQO state would be restored and the
external changes would be lost.