Improve mutation behavior for RelNodes
The client no longer needs to call mutate() on all input nodes of a mutated node
themself. Instead, the mutation of all child nodes is handled as part of the
mutation of the parent node automatically. This works thanks to the as_root switch,
which is now mandatory a mandatory argument of all mutate() implementations.