Fix regression introduced in ace5ea3246ccdafe1dfe46a7e5f2d0777af9c61c
That commit differentiated between the original query object passed into
the execute_query() method and the object that will get executed in the
end. However, the original object was still stored in the query cache.
After the change, the cache could receive instances of SqlQuery instead.
In turn, this would break the JSON representation of the cache, since
json.dump only supports primitive objects.