Fix PG timeout executor not cancelling the query on the server side
As it turns out, killing the Python process that maintains the psycopg
connection does not result in psycopg actually cancelling the query on
the PG server side. Instead, this query would finish normally and then
be rolled back
(see https://www.psycopg.org/articles/2014/07/20/cancelling-postgresql-statements-python/)
This is obviously not what we want since the query would still eat our
resources.

As a solution, we now have a dedicated timeout watchdog that issues
pg_cancel_backend() commands for all connections that run into a
timeout.